Don't require a Python test for now.
